requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai>=1.40.0
python-docx==1.1.2
python-multipart==0.0.9
